added user dropdown

+++ src/components/Stripe/Stripe.js
@@ -3,20 +3,25 @@ import "./Stripe.css";
 import StripeCheckout from "react-stripe-checkout";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteCartItems } from "../../actions/index";
-import { deleteCartItems as deleteCartFirebase,setOrderHistory } from "../../firebase/firebase";
+import {
+  deleteCartItems as deleteCartFirebase,
+  setOrderHistory,
+} from "../../firebase/firebase";
 
 export default function Stripe({ price, setIsOrderCompleted }) {
   const newPrice = price * 100;
   const key = process.env.REACT_APP_STRIPE;
   const dispatch = useDispatch();
   const userId = useSelector((state) => state.authUserState.authUserId);
-  const orderItems = useSelector((state)=> state.cartState.cart)
+  const orderItems = useSelector((state) => state.cartState.cart);
 
   const onToken = () => {
-    setOrderHistory(userId,orderItems)
     dispatch(deleteCartItems());
-    deleteCartFirebase(userId);
     setIsOrderCompleted(true);
+    if (userId) {
+      setOrderHistory(userId, orderItems);
+      deleteCartFirebase(userId);
+    }
   };
   return (
     <StripeCheckout
